Move all conda env code to its own file
